read train samples

+++ lib/Parameters.py
@@ -82,9 +82,9 @@ def deal_args_active(my_argv):
         elif opt in ("-a", "--active_size"):
             a = int(arg)
         elif opt in ("-u", "--threshold_up"):
-            u = int(arg)
+            u = float(arg)
         elif opt in ("-l", "--threshold_low"):
-            l = int(arg)
+            l = float(arg)
     print 'settings: v=%s, d=%s, n1=%d, n0=%d, b=%d, e=%d, t=%d, z=%d, m=%s, a=%d, u=%f, l=%f' % (
         v, d, n1, n0, b, e, t, z, m, a, u, l)
     return v, d, n1, n0, b, e, t, z, m, a, u, l
